root_files/cpu_requirements.txt
# THIS requirements file was created  for https://huggingface.co/spaces/GIZ/cpu_tracs on 20.02.2024
certifi==2024.2.2
pytz==2024.1
tzdata==2024.1
regex==2023.12.25
jsonschema-specifications==2023.12.1
fsspec==2023.10.0
setuptools==65.5.1
cattrs==23.2.3
attrs==23.2.0
packaging==23.2
pip==23.0.1
pyarrow==15.0.0
rich==13.7.0
nvidia-nvjitlink-cu12==12.3.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
nvidia-cublas-cu12==12.1.3.1
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
more-itertools==10.2.0
pillow==9.1.1
nvidia-cudnn-cu12==8.9.2.26
tenacity==8.2.3
click==8.1.7
thinc==8.0.17
importlib-metadata==7.0.1
smart-open==6.4.0
tornado==6.4
multidict==6.0.5
pyyaml==6.0.1
inflect==6.0.1
plotly==5.19.0
cachetools==5.3.2
tzlocal==5.2
smmap==5.0.1
tqdm==4.66.2
fonttools==4.49.0
transformers==4.36.2
protobuf==4.25.3
jsonschema==4.21.1
typing-extensions==4.9.0
anyio==4.3.0
altair==4.2.2
platformdirs==4.2.0
gitdb==4.0.11
async-timeout==4.0.3
prompthub-py==4.0.0
watchdog==4.0.0
zipp==3.17.0
filelock==3.13.1
aiohttp==3.9.3
python-decouple==3.8
nltk==3.7
idna==3.6
matplotlib==3.5.1
posthog==3.4.1
xxhash==3.4.1
charset-normalizer==3.3.2
langcodes==3.3.0
threadpoolctl==3.3.0
networkx==3.2.1
xlsxwriter==3.2.0
en-core-web-sm==3.2.0
spacy==3.2.0
gitpython==3.1.42
jinja2==3.1.3
pyparsing==3.1.1
spacy-legacy==3.0.12
preshed==3.0.9
markdown-it-py==3.0.0
requests==2.31.0
nvidia-nccl-cu12==2.18.1
pygments==2.17.2
datasets==2.17.1
python-dateutil==2.8.2
srsly==2.4.8
sentence-transformers==2.3.1
backoff==2.2.1
urllib3==2.2.1
pandas==2.2.0
markupsafe==2.1.5
torch==2.1.0
triton==2.1.0
catalogue==2.0.10
cymem==2.0.8
streamlit==1.31.0
numpy==1.26.4
farm-haystack==1.24.0
pymupdf==1.23.24
pymupdfb==1.23.22
pdf2image==1.17.0
six==1.16.0
scipy==1.12.0
sympy==1.12
yarl==1.9.4
pydantic==1.8.2
sseclient-py==1.8.0
blinker==1.7.0
monotonic==1.6
kiwisolver==1.4.5
appdirs==1.4.4
url-normalize==1.4.3
frozenlist==1.4.1
joblib==1.3.2
scikit-learn==1.3.2
aiosignal==1.3.1
mpmath==1.3.0
sniffio==1.3.0
exceptiongroup==1.2.0
summa==1.2.0
murmurhash==1.0.10
langdetect==1.0.9
boilerpy3==1.0.7
spacy-loggers==1.0.5
httpcore==1.0.3
setfit==1.0.3
utils==1.0.1
multiprocess==0.70.16
wheel==0.41.1
referencing==0.33.0
httpx==0.26.0
validators==0.22.0
huggingface-hub==0.20.3
responses==0.18.0
rpds-py==0.18.0
tokenizers==0.15.2
h11==0.14.0
cycler==0.12.1
toolz==0.12.1
seaborn==0.11.2
pathy==0.11.0
toml==0.10.2
wasabi==0.10.1
requests-cache==0.9.8
quantulum3==0.9.0
pydeck==0.8.1b0
blis==0.7.11
docopt==0.6.2
pyarrow-hotfix==0.6
tiktoken==0.6.0
num2words==0.5.13
events==0.5
safetensors==0.4.2
typer==0.4.2
evaluate==0.4.1
entrypoints==0.4
pytesseract==0.3.10
dill==0.3.8
streamlit-aggrid==0.3.4.post3
lazy-imports==0.3.1
rank-bm25==0.2.2
sentencepiece==0.2.0
mdurl==0.1.2
pathlib-abc==0.1.1
